parser/htmlparser/tests/mochitest/test_bug1752498.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title><!-- TODO: insert title here --></title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    add_task(async function test() {
      let string = "<p>Not a <<b>well-formed</b> xml string</p>";
      // Line and column are 1-based.
      let line = (string.match(/\r\n|\r|\n/g)?.length || 0) + 1;
      let column = string.search("<b>") + 1;
      let expectedError = `XML Parsing Error: not well-formed
Location: ${location.href}
Line Number ${line}, Column ${column}:${string}
${"^".padStart(column, "-")}`;

      let error = new window.DOMParser()
        .parseFromString(string, "text/xml")
        .getElementsByTagName("parsererror")[0].textContent;

      is(error, expectedError, "Check that parsererror contains the right data.");
    });
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
